refactor: Eliminar componentes sin desarrollar y preparar para integración de pago con visualizar experiencia
- Eliminados componentes no desarrollados que no aportan funcionalidad en el flujo actual.
- Organizado y limpiado el código para preparar la integración entre la funcionalidad de pago y la visualización de la experiencia.
- Ajustes generales para mejorar la estructura del proyecto y facilitar la futura integración.
