+++ Front-End/src/app/app-routing.module.ts
@@ -1,19 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { VisualizarExperienciaComponent } from './paginas/visualizar-experiencia/visualizar-experiencia.component';
-import { ContenidoPagoComponent } from './paginas/pago/contenido-pago/contenido-pago.component';
-import { AnuncioComponent } from './paginas/anuncio/anuncio.component';
-import { PerfilComponent } from './paginas/perfil/perfil.component';
-import { LoginComponent } from './paginas/login/login.component';
-import { ContenidoReprogramacionComponent } from './paginas/reprogramacion/contenido-reprogramacion/contenido-reprogramacion.component'; // ✅ NUEVO
+import { ContenidoPagoComponent } from './components/pago/contenido-pago/contenido-pago.component';
+import { DisponibilidadComponent } from './components/disponibilidad/disponibilidad.component';
 
 const routes: Routes = [
-  { path: '', component: VisualizarExperienciaComponent },
+  { path: '', component: DisponibilidadComponent },
   { path: 'pago', component: ContenidoPagoComponent },
-  { path: 'anuncio', component: AnuncioComponent },
-  { path: 'perfil', component: PerfilComponent },
-  { path: 'login', component: LoginComponent },
-  { path: 'reprogramacion', component: ContenidoReprogramacionComponent } // ✅ CORREGIDO
 ];
 
 @NgModule({
