+++ Front-End/src/app/app.component.html
@@ -1,2 +1,1 @@
-
 <router-outlet></router-outlet>
