+++ Front-End/src/app/components/disponibilidad/disponibilidad.component.spec.ts
@@ -1,18 +1,18 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { LoginComponent } from './login.component';
+import { DisponibilidadComponent } from './disponibilidad.component';
 
-describe('LoginComponent', () => {
-  let component: LoginComponent;
-  let fixture: ComponentFixture<LoginComponent>;
+describe('DisponibilidadComponent', () => {
+  let component: DisponibilidadComponent;
+  let fixture: ComponentFixture<DisponibilidadComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ LoginComponent ]
+      declarations: [ DisponibilidadComponent ]
     })
     .compileComponents();
 
-    fixture = TestBed.createComponent(LoginComponent);
+    fixture = TestBed.createComponent(DisponibilidadComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
