+++ Front-End/src/app/components/pago/contenido-pago/contenido-pago.component.css
