+++ Front-End/src/app/components/pago/pago.component.ts
