--- Front-End/src/app/paginas/anuncio/anuncio.component.html
@@ -1,1 +1,0 @@
-<p>anuncio works!</p>
