--- Front-End/src/app/paginas/anuncio/anuncio.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-anuncio',
-  templateUrl: './anuncio.component.html',
-  styleUrls: ['./anuncio.component.css']
-})
-export class AnuncioComponent {
-
-}
