--- Front-End/src/app/paginas/login/login.component.html
@@ -1,1 +1,0 @@
-<p>login works!</p>
