--- Front-End/src/app/paginas/login/login.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-login',
-  templateUrl: './login.component.html',
-  styleUrls: ['./login.component.css']
-})
-export class LoginComponent {
-
-}
