--- Front-End/src/app/paginas/navbar/navbar.component.html
@@ -1,12 +1,0 @@
-<div class="container py-5">
-  <!-- HEADER -->
-  <header class="d-flex justify-content-between align-items-center border-bottom pb-3 mb-4">
-    <img src="assets/expedia-logo.png" alt="Expedia" style="height: 40px;" />
-    <nav class="d-flex gap-3 align-items-center">
-      <a routerLink="/anuncio" class="text-decoration-none fw-bold text-dark">Anunciar una propiedad</a>
-      <a href="#" class="text-decoration-none fw-bold text-dark">Atención al cliente</a>
-      <a routerLink="/perfil" class="text-decoration-none fw-bold text-dark">Mis viajes</a>
-      <a routerLink="/login" class="btn btn-outline-dark">Iniciar Sesión</a>
-    </nav>
-  </header>
-</div>
