--- Front-End/src/app/paginas/pago/pago.component.html
@@ -1,152 +1,0 @@
-<div class="container py-5">
-  <!-- HEADER -->
-  <header class="d-flex justify-content-between align-items-center border-bottom pb-3 mb-4">
-    <img src="assets/expedia-logo.png" alt="Expedia" style="height: 40px;" />
-    <nav class="d-flex gap-3 align-items-center">
-      <a routerLink="/anuncio" class="text-decoration-none fw-bold text-dark">Anunciar una propiedad</a>
-      <a href="#" class="text-decoration-none fw-bold text-dark">Atención al cliente</a>
-      <a routerLink="/perfil" class="text-decoration-none fw-bold text-dark">Mis viajes</a>
-      <a routerLink="/login" class="btn btn-outline-dark">Iniciar Sesión</a>
-    </nav>
-  </header>
-
-  <form *ngIf="formulario" [formGroup]="formulario" (ngSubmit)="reservar()" class="card shadow p-4">
-
-    <!-- ENCARGADO DE LA RESERVA -->
-    <section class="mb-5">
-      <h3 class="mb-4"><i class="bi bi-person-badge-fill me-2"></i>Encargado de la reserva</h3>
-      <div [formGroup]="encargadoGroup" class="border rounded p-3">
-        <div class="mb-3">
-          <label>Nombre completo:</label>
-          <input type="text" class="form-control" formControlName="nombre" placeholder="Ej: Juan Pérez"
-            [class.is-invalid]="campoInvalido('encargado.nombre')" />
-          <div class="invalid-feedback">Solo letras. Obligatorio.</div>
-        </div>
-
-        <div class="mb-3">
-          <label>Correo electrónico:</label>
-          <input type="email" class="form-control" formControlName="correo" placeholder="correo@ejemplo.com"
-            [class.is-invalid]="campoInvalido('encargado.correo')" />
-          <div class="invalid-feedback">Correo no válido.</div>
-        </div>
-
-        <div class="mb-3">
-          <label>Teléfono:</label>
-          <input type="text" class="form-control" formControlName="telefono" placeholder="912345678" inputmode="numeric"
-            pattern="[0-9]*" (keydown)="soloNumeros($event)" [class.is-invalid]="campoInvalido('encargado.telefono')" />
-        </div>
-      </div>
-      <!-- RESUMEN DE PERSONAS -->
-      <section class="mt-4">
-        <h5 class="fw-semibold text-primary">
-          <i class="bi bi-people-fill"></i> Personas que viajan:
-        </h5>
-        <ul class="list-group">
-          <li class="list-group-item">
-            <i class="bi bi-person-fill text-dark me-2 fs-5"></i>
-            Adultos: {{ datosRecibidos.adultos }}
-          </li>
-          <li class="list-group-item">
-            <i class="bi bi-person-circle text-dark me-2 fs-5"></i>
-            Niños: {{ datosRecibidos.ninos }}
-          </li>
-        </ul>
-      </section>
-
-
-    </section>
-
-    <!-- MÉTODO DE PAGO -->
-    <section class="mb-5">
-      <h3 class="mb-4"><i class="bi bi-credit-card me-2"></i>Pago</h3>
-
-      <!-- OPCIONES -->
-      <div class="d-flex flex-wrap gap-3 mb-4">
-        <div class="p-3 border rounded text-center flex-fill"
-          [ngClass]="{'border-primary': metodoSeleccionado === 'credito'}" style="cursor: pointer; min-width: 220px"
-          (click)="seleccionarMetodo('credito')">
-          <i class="bi bi-credit-card-fill fs-4 d-block mb-2"></i>
-          Tarjeta de Crédito
-        </div>
-        <div class="p-3 border rounded text-center flex-fill"
-          [ngClass]="{'border-primary': metodoSeleccionado === 'debito'}" style="cursor: pointer; min-width: 220px"
-          (click)="seleccionarMetodo('debito')">
-          <i class="bi bi-credit-card-fill fs-4 d-block mb-2"></i>
-          Tarjeta de Débito
-        </div>
-        <div class="p-3 border rounded text-center flex-fill"
-          [ngClass]="{'border-primary': metodoSeleccionado === 'paypal'}" style="cursor: pointer; min-width: 220px"
-          (click)="seleccionarMetodo('paypal')">
-          <i class="bi bi-paypal fs-4 d-block mb-2"></i>
-          PayPal
-        </div>
-      </div>
-
-      <!-- FORMULARIO TARJETA -->
-      <div *ngIf="mostrarFormulario && (metodoSeleccionado === 'credito' || metodoSeleccionado === 'debito')"
-        [formGroup]="pagoGroup" [@fadeIn]>
-        <div class="mb-3">
-          <label>Nombre en la tarjeta:</label>
-          <input type="text" class="form-control" formControlName="nombreTarjeta" placeholder="Ej: Juan Pérez"
-            [class.is-invalid]="campoInvalido('pago.nombreTarjeta')" />
-          <div class="invalid-feedback">Solo letras. Campo obligatorio.</div>
-        </div>
-        <div class="mb-3">
-          <label>Número de tarjeta:</label>
-          <input type="text" class="form-control" formControlName="numeroTarjeta" placeholder="1234 5678 9012 3456"
-            maxlength="16" inputmode="numeric" pattern="[0-9]*" (keydown)="soloNumeros($event)"
-            [class.is-invalid]="campoInvalido('pago.numeroTarjeta')" />
-          <div class="invalid-feedback">Debe tener 16 dígitos numéricos.</div>
-        </div>
-        <div class="row">
-          <div class="col-md-6 mb-3">
-            <label>Fecha de vencimiento (MM/AA):</label>
-            <input type="text" class="form-control" formControlName="vencimiento" placeholder="08/26"
-              inputmode="numeric" (keydown)="soloNumerosYSlash($event)" maxlength="5"
-              [class.is-invalid]="campoInvalido('pago.vencimiento')" />
-            <div class="invalid-feedback">Formato inválido. Usa MM/AA.</div>
-          </div>
-          <div class="col-md-6 mb-3">
-            <label>Código de seguridad (CVV):</label>
-            <input type="text" class="form-control" formControlName="cvv" placeholder="123" maxlength="3"
-              inputmode="numeric" pattern="[0-9]*" (keydown)="soloNumeros($event)"
-              [class.is-invalid]="campoInvalido('pago.cvv')" />
-            <div class="invalid-feedback">Debe tener 3 dígitos.</div>
-          </div>
-        </div>
-      </div>
-      <!-- FORMULARIO PAYPAL -->
-      <div *ngIf="mostrarFormulario && metodoSeleccionado === 'paypal'" [formGroup]="paypalGroup" [@fadeIn]>
-        <div class="mb-3">
-          <label>Correo electrónico de PayPal:</label>
-          <input type="email" class="form-control" formControlName="correoPaypal" placeholder="usuario@paypal.com"
-            [class.is-invalid]="campoInvalido('paypal.correoPaypal')" />
-          <div class="invalid-feedback">Correo inválido.</div>
-        </div>
-        <div class="mb-3">
-          <label>Contraseña de PayPal:</label>
-          <input type="password" class="form-control" formControlName="contrasenaPaypal"
-            [class.is-invalid]="campoInvalido('paypal.contrasenaPaypal')" />
-          <div class="invalid-feedback">Campo obligatorio (mínimo 6 caracteres).</div>
-        </div>
-      </div>
-    </section>
-
-    <!-- RESUMEN -->
-    <section class="mb-4">
-      <h3 class="mb-3"><i class="bi bi-clipboard-check me-2"></i>Resumen de la experiencia</h3>
-      <ul class="list-group">
-        <li class="list-group-item">Ubicación: {{ datosRecibidos.ciudad }}</li>
-        <li class="list-group-item">Experiencia: {{ datosRecibidos.titulo }}</li>
-        <li class="list-group-item">Fecha: {{ datosRecibidos.fecha }}</li>
-        <li class="list-group-item">Hora: {{ datosRecibidos.hora }}</li>
-        <li class="list-group-item fw-bold">Total a pagar: CLP {{ datosRecibidos.total }}</li>
-      </ul>
-    </section>
-
-    <!-- BOTÓN -->
-    <button type="submit" class="btn btn-primary w-100 fs-5" [disabled]="formulario.invalid">
-      <i class="bi bi-check-circle-fill me-2"></i>Reservar
-    </button>
-  </form>
-</div> <!-- Fin del container -->
