--- Front-End/src/app/paginas/perfil/perfil.component.html
@@ -1,1 +1,0 @@
-<p>perfil works!</p>
