--- Front-End/src/app/paginas/perfil/perfil.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-perfil',
-  templateUrl: './perfil.component.html',
-  styleUrls: ['./perfil.component.css']
-})
-export class PerfilComponent {
-
-}
