--- Front-End/src/app/paginas/visualizar-experiencia/visualizar-experiencia.component.css
@@ -1,260 +1,0 @@
-.contenedor {
-  font-family: Arial, sans-serif;
-  padding: 20px;
-  max-width: 1200px;
-  margin: auto;
-  background-color: #fff;
-  color: #333;
-}
-
-.encabezado {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  border-bottom: 1px solid #ccc;
-  padding-bottom: 10px;
-}
-
-.logo {
-  height: 40px;
-}
-
-.encabezado nav a {
-  margin-right: 20px;
-  color: #333;
-  text-decoration: none;
-  font-weight: bold;
-}
-
-.btn-iniciar {
-  background-color: #fff;
-  border: 2px solid #333;
-  padding: 6px 12px;
-  cursor: pointer;
-  font-weight: bold;
-}
-
-.experiencia h2 {
-  margin-top: 30px;
-  font-size: 24px;
-  line-height: 1.4;
-}
-
-.experiencia h2 span {
-  font-weight: bold;
-  font-size: 28px;
-  display: block;
-  color: #004f95;
-}
-
-/* ✅ GALERÍA ESTILO EXPEDIA */
-.galeria-expedia {
-  display: flex;
-  gap: 10px;
-  margin-top: 20px;
-  align-items: flex-start;
-}
-
-.imagen-grande {
-  flex: 2;
-}
-
-.imagen-grande img {
-  width: 100%;
-  height: 400px;
-  object-fit: cover;
-  border-radius: 8px;
-}
-
-.imagenes-derecha {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.imagenes-superiores {
-  display: flex;
-  gap: 10px;
-}
-
-.imagenes-superiores img {
-  width: 50%;
-  height: 95px;
-  object-fit: cover;
-  border-radius: 8px;
-}
-
-.imagen-inferior img {
-  width: 100%;
-  height: 290px;
-  object-fit: cover;
-  border-radius: 8px;
-}
-
-
-.informacion {
-  display: flex;
-  flex-wrap: wrap;
-  margin-top: 30px;
-  gap: 30px;
-}
-
-.datos {
-  flex: 2;
-}
-
-.datos ul {
-  list-style: none;
-  padding: 0;
-}
-
-.datos ul li {
-  margin-bottom: 8px;
-}
-
-.fechas {
-  margin: 15px 0;
-}
-
-.fechas button {
-  padding: 10px 15px;
-  margin-right: 10px;
-  background-color: #f9f9f9;
-  border: 1px solid #ccc;
-  border-radius: 6px;
-  font-weight: bold;
-  cursor: pointer;
-  min-width: 140px;
-  transition: background-color 0.2s ease, color 0.2s ease, border-color 0.2s ease;
-}
-
-.fechas button:hover {
-  background-color: #0071c2;
-  border-color: #0071c2;
-  color: white;
-}
-
-.fechas button.seleccionada {
-  background-color: #0071c2;
-  border-color: #0071c2;
-  color: white;
-}
-
-.resumen {
-  margin-top: 20px;
-  border-top: 1px solid #ccc;
-  padding-top: 15px;
-}
-
-.resumen p {
-  margin: 5px 0;
-}
-
-.total {
-  font-size: 18px;
-  font-weight: bold;
-  color: #004f95;
-}
-
-.btn-reservar {
-  margin-top: 10px;
-  background-color: #0071c2;
-  color: white;
-  padding: 10px 15px;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-.btn-reservar:hover {
-  background-color: #005fa3;
-}
-
-.mapa {
-  flex: 1;
-}
-
-.mapa iframe {
-  width: 100%;
-  height: 250px;
-  border: none;
-  border-radius: 6px;
-}
-
-.rating {
-  margin-top: 10px;
-  background-color: #e6f4ff;
-  padding: 10px;
-  border-radius: 6px;
-  text-align: center;
-  font-weight: bold;
-}
-
-.incluye,
-.descripcion {
-  margin-top: 40px;
-}
-
-.incluye ul,
-.descripcion p {
-  padding-left: 0;
-  list-style: none;
-}
-
-.incluye li {
-  margin-bottom: 6px;
-}
-
-.descripcion p {
-  line-height: 1.6;
-  margin-top: 10px;
-}
-
-/* --- NUEVO: Estilo para la sección de disponibilidad con resumen --- */
-
-.resumen-box {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 15px;
-  padding: 20px;
-  border: 1px solid #aaa;
-  border-radius: 8px;
-  background-color: #fff;
-  gap: 20px;
-  flex-wrap: wrap;
-}
-
-.resumen-left,
-.resumen-mid,
-.resumen-right {
-  flex: 1;
-  min-width: 200px;
-}
-
-.resumen-mid {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.selector {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  margin-bottom: 5px;
-}
-
-.selector button {
-  padding: 8px 12px;
-  border: 1px solid #ccc;
-  border-radius: 6px;
-  background-color: #fff;
-  cursor: pointer;
-  transition: background-color 0.2s ease, border-color 0.2s ease;
-}
-
-.selector button:hover {
-  background-color: #e6f4ff;
-  border-color: #0071c2;
-}
